+++ src/headers/No.hpp
@@ -6,7 +6,7 @@ class No {
     float desconto;
     int preco;
     int tempo;
-    int custo_acumulado;
+    float custo_acumulado;
 
     public:
         No();
@@ -14,7 +14,7 @@ class No {
         void set_preco_e_tempo(int p, int t);
         void get_desconto_tempo_e_preco();
         int get_preco();
-        int get_desconto();
+        float get_desconto();
         int get_tempo();
         
 };
